+++ code-challenge/src/main/java/com/megamaker/codechallenge/domain/entity/UserBadge.java
@@ -1,4 +1,4 @@
-package com.megamaker.codechallenge.entity;
+package com.megamaker.codechallenge.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
